requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
mysql-connector-python==8.2.0
pydantic==2.4.2
